Do not autoCancel paused notification, hides status one in Pie
Apparently Android Pie now automatically cancels the notification
after the code is executed, rather than before. This causes not
the "paused" notification to be cancelled, but the "running" one.

Fix this by introducing a new "RESUME" command that manually cancels
the "paused" notification and then falls through to start.

Fixes #320